TestMergeProject(준혁 보완)  : 장바구니-주문완료창 연습1.3차 저장  db저장까지는 해결  모달창 안뜨는에러발생(변형되었으니 받지마세요 )

+++ src/main/java/com/example/joinair/dto/ORDER_DETAIL.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class ORDER_DETAIL {
-    private int OD_Num; // 주문상세번호    (ORDER_DETAIL의 기본키, 자동1증가생성 시퀀스 만들어줘야 하고 DB에 적용해줘야함 . )
+    private Long OD_Num; // 주문상세번호    (ORDER_DETAIL의 기본키, 자동1증가생성 시퀀스 만들어줘야 하고 DB에 적용해줘야함 . )
     private String OD_Pro_Name; // 주문한 상품명    ( DB컬럼 생성 필요함)
     private int OD_Pro_Price; // 주문한 상품명의 단가  (DB컬럼 생성 필요함)
     private int OD_Qua; // 수량             (session에 담긴 Quantity)
